refactor(lotto): 패키지 구조 변경
Logger 클래스를 util 패키지로 이동
ErrorMessage 클래스를 validator 패키지로 이동
LotteryPublisher, Lotto, LottoRank, LottoResult, LottoWinningDiscriminator, User 클래스를 core 패키지로 이동
