+++ src/test/java/lotto/UserTest.java
@@ -1,15 +1,16 @@
 package lotto;
 
-import static lotto.ErrorMessage.PURCHASE_AMOUNT_NOT_NUMBER_INPUT_ERROR;
-import static lotto.ErrorMessage.PURCHASE_AMOUNT_NOT_DIVIDED_BY_1000_ERROR;
-import static lotto.ErrorMessage.WINNING_NUMBER_NOT_NUMBER_INPUT_ERROR;
-import static lotto.ErrorMessage.INVALID_LOTTO_NUMBER_RANGE_ERROR;
-import static lotto.ErrorMessage.INVALID_LOTTO_NUMBER_SIZE_ERROR;
-import static lotto.ErrorMessage.BONUS_NUMBER_NOT_NUMBER_INPUT_ERROR;
-import static lotto.ErrorMessage.LOTTO_NUMBER_DUPLICATE_ERROR;
-import static lotto.ErrorMessage.BONUS_NUMBER_EXIST_IN_WINNING_NUMBER_ERROR;
+import static lotto.validator.ErrorMessage.PURCHASE_AMOUNT_NOT_NUMBER_INPUT_ERROR;
+import static lotto.validator.ErrorMessage.PURCHASE_AMOUNT_NOT_DIVIDED_BY_1000_ERROR;
+import static lotto.validator.ErrorMessage.WINNING_NUMBER_NOT_NUMBER_INPUT_ERROR;
+import static lotto.validator.ErrorMessage.INVALID_LOTTO_NUMBER_RANGE_ERROR;
+import static lotto.validator.ErrorMessage.INVALID_LOTTO_NUMBER_SIZE_ERROR;
+import static lotto.validator.ErrorMessage.BONUS_NUMBER_NOT_NUMBER_INPUT_ERROR;
+import static lotto.validator.ErrorMessage.LOTTO_NUMBER_DUPLICATE_ERROR;
+import static lotto.validator.ErrorMessage.BONUS_NUMBER_EXIST_IN_WINNING_NUMBER_ERROR;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.core.User;
 import lotto.input.MockInputReader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
